requirements.txt
python-dotenv
watchdog
cryptography
